order-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
azure-storage-queue==12.9.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
